src/requirements.txt
absl-py==0.7.0
astor==0.7.1
astunparse==1.6.3
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
editdistance==0.5.2
future==0.18.2
gast==0.2.2
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
joblib==0.16.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.0.1
matplotlib==3.3.0
networkx==2.4
numpy==1.16.1
oauthlib==3.1.0
opencv-python==4.0.0.21
opt-einsum==3.3.0
pandas==1.0.5
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
six==1.12.0
sklearn==0.0
soupsieve==2.0.1
tensorboard==2.0.2
tensorboard-plugin-wit==1.7.0
tensorboardX==1.9
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.7.24
torch==1.5.1
torchvision==0.6.1
tqdm==4.48.0
urllib3==1.25.10
Werkzeug==0.14.1
wrapt==1.12.1
zipp==3.1.0
